fix: averageAge 타입 변경에 따른 dto 타입 변경

+++ project/back/health/src/main/java/com/ssafy/health/domain/crew/dto/response/CrewDetailResponseDto.java
@@ -16,7 +16,7 @@ public class CrewDetailResponseDto {
     private Long crewRanking;
     private Long totalBattlesCount;
     private Long winCount;
-    private Float averageAge;
+    private Integer averageAge;
     private Float basicScore;
     private Float activityScore;
     private CrewRole role;
